refactor(backup): move cold_backup_metadata out to standalone file
