+++ src/replica/CMakeLists.txt
@@ -9,7 +9,8 @@ set(DUPLICATION_SRC
         duplication/mutation_batch.cpp
 )
 
-set(BACKUP_SRC backup/replica_backup_manager.cpp)
+set(BACKUP_SRC backup/replica_backup_manager.cpp
+               backup/cold_backup_context.cpp)
 
 set(BULK_LOAD_SRC bulk_load/replica_bulk_loader.cpp)
